Rework test user creation mechanism (#54)
This PR stops WingedKeys from removing all user data from the db on deploy of all non-prod environments, and instead only creates a default admin user if one is not already present. It also adds Log In and Create User components to the navbar (the latter exclusively for admin users) so that user creation is a bit more accessible in non-prod environments.